Add in auth support and examples
This adds in auth support for services that support the
capabilities endpoint.  This is where the auth information about
supported methods is exchanged.  The framework is flexible enough
to allow people to add auth to any call in pyvo though, by
configuring a session to be used by those calls.

Also add some examples in of how to use these functions with
services that already do auth and are available on the internet.